# api/openapi.yaml
openapi: 3.2.0
info:
  title: ABCMovies
  description: |
    This is the API Documentation for ABCMovies, a program that allows people to
    download and stream videos from many different streaming services.

    This specification aims to align with:
      - ISO 3166-1 alpha-2 for countries
      - ISO 639-1 for languages
      - ISO 8601 for durations and dates
      - RFC 3339 for date/time formats
      - schema.org for resource type identification
  version: 0.0.0-alpha
servers:
  - url: http://localhost:80/api/v1alpha
    description: Local development
  - url: https://api.abcmovies.example/v1alpha
    description: Production
...
paths:
...
  /services:
    get:
      tags:
        - Services
      summary: List all streaming services
      description: Retrieve all streaming services offered by the API.
      operationId: getServices
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of streaming services
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Service'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /services/{serviceTag}:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
    get:
      tags:
        - Services
      summary: Get a streaming service by tag
      description: Retrieve metadata for a specific streaming service.
      operationId: getServiceByTag
      responses:
        '200':
          description: A streaming service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Movies ---------------------------------------------------------------
  /services/{serviceTag}/movies:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
    get:
      tags:
        - Movies
      summary: List movies on a service
      description: Retrieve all movies available on the given streaming service.
      operationId: getMovies
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of movies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Movie'
        '400':
          description: Invalid service tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /services/{serviceTag}/movies/{movieId}:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
      - $ref: '#/components/parameters/MovieId'
    get:
      tags:
        - Movies
      summary: Get a movie
      description: Retrieve a specific movie by its identifier on the streaming service.
      operationId: getMovieById
      responses:
        '200':
          description: A movie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '400':
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Movie Streams --------------------------------------------------------
  /services/{serviceTag}/movies/{movieId}/streams:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
      - $ref: '#/components/parameters/MovieId'
    get:
      tags:
        - Movies
        - Streams
      summary: List streams for a movie
      description: |
        Retrieve all available video stream manifests (DASH, HLS, etc.)
        for a movie. The listing returns metadata only; fetch the individual
        file endpoint to get the raw manifest bytes.
      operationId: getMovieStreams
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of video stream metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Stream'
        '400':
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Movie not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /services/{serviceTag}/series:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
    get:
      tags:
        - Series
      summary: List TV series on a service
      description: Retrieve all TV series available on the given streaming service.
      operationId: getSeries
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of TV series
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Series'
        '400':
          description: Invalid service tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Service not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /services/{serviceTag}/series/{seriesId}:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
      - $ref: '#/components/parameters/SeriesId'
    get:
      tags:
        - Series
      summary: Get a TV series
      description: Retrieve a specific TV series by its identifier on the streaming service.
      operationId: getSeriesById
      responses:
        '200':
          description: A TV series
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Series'
        '400':
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Series not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Series Images ---------------------------------------------------------
...
  /services/{serviceTag}/series/{seriesId}/seasons:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
      - $ref: '#/components/parameters/SeriesId'
    get:
      tags:
        - Seasons
      summary: List seasons for a series
      description: Retrieve all seasons for a given TV series.
      operationId: getSeasons
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of seasons
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Season'
        '400':
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Series not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /services/{serviceTag}/series/{seriesId}/seasons/{seasonId}:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
      - $ref: '#/components/parameters/SeriesId'
      - $ref: '#/components/parameters/SeasonId'
    get:
      tags:
        - Seasons
      summary: Get a season
      description: Retrieve a specific season by its identifier.
      operationId: getSeasonById
      responses:
        '200':
          description: A season
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Season'
        '400':
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Season not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Season Images ---------------------------------------------------------
...
  /services/{serviceTag}/series/{seriesId}/seasons/{seasonId}/episodes:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
      - $ref: '#/components/parameters/SeriesId'
      - $ref: '#/components/parameters/SeasonId'
    get:
      tags:
        - Episodes
      summary: List episodes in a season
      description: Retrieve all episodes for a given season.
      operationId: getEpisodes
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of episodes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Episode'
        '400':
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Season not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /services/{serviceTag}/series/{seriesId}/seasons/{seasonId}/episodes/{episodeId}:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
      - $ref: '#/components/parameters/SeriesId'
      - $ref: '#/components/parameters/SeasonId'
      - $ref: '#/components/parameters/EpisodeId'
    get:
      tags:
        - Episodes
      summary: Get an episode
      description: Retrieve a specific episode by its identifier.
      operationId: getEpisodeById
      responses:
        '200':
          description: An episode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Episode'
        '400':
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Episode not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # -- Episode Images --------------------------------------------------------
...
  /services/{serviceTag}/series/{seriesId}/seasons/{seasonId}/episodes/{episodeId}/streams:
    parameters:
      - $ref: '#/components/parameters/ServiceTag'
      - $ref: '#/components/parameters/SeriesId'
      - $ref: '#/components/parameters/SeasonId'
      - $ref: '#/components/parameters/EpisodeId'
    get:
      tags:
        - Episodes
        - Streams
      summary: List streams for an episode
      description: |
        Retrieve all available video stream manifests for an episode.
        The listing returns metadata only; fetch the individual file
        endpoint to get the raw manifest bytes.
      operationId: getEpisodeStreams
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: A page of video stream metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_Stream'
        '400':
          description: Invalid path parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Episode not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
...
    ServiceTag:
      name: serviceTag
      in: path
      description: |
        A short, unique tag identifying the streaming service.
        Example: `CRAV`, `NFLX`, `DSNP`.
      required: true
      schema:
        type: string
        pattern: '^[A-Z]{2,8}$'
        examples:
          - CRAV
    MovieId:
      name: movieId
      in: path
      description: |
        The movie's identifier on the streaming service.
        Preferably a readable slug, not a numeric ID, when possible.
      required: true
      schema:
        type: string
        examples:
          - inception-2010
    SeriesId:
      name: seriesId
      in: path
      description: |
        The TV series's identifier on the streaming service.
        Preferably a readable slug, not a numeric ID, when possible.
      required: true
      schema:
        type: string
        examples:
          - hells-kitchen-52597
    SeasonId:
      name: seasonId
      in: path
      description: |
        Identifier for the season. Prefer a human-readable slug
        (e.g. `season-1`) or a descriptive name when available.
      required: true
      schema:
        type: string
        examples:
          - season-1
...
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return per page.
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      description: Number of results to skip for offset-based pagination.
      schema:
        type: integer
        default: 0
        minimum: 0
...
  schemas:
...
    Error:
      type: object
      description: An error response.
      additionalProperties: false
      properties:
        code:
          type: string
          description: Machine-readable error code.
          examples:
            - NOT_FOUND
        message:
          type: string
          description: Human-readable error description.
          examples:
            - The requested resource was not found.
      required:
        - code
        - message

    # -- Pagination ----------------------------------------------------------
    Page_Service:
      type: object
      description: Paginated list of streaming services.
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Service'
        total:
          type: integer
          description: Total number of matching items across all pages.
          minimum: 0
        limit:
          type: integer
          description: Maximum items per page (as requested).
          minimum: 1
        offset:
          type: integer
          description: Zero-based offset of the first item in this page.
          minimum: 0
      required:
        - items
        - total
        - limit
        - offset
    Page_Movie:
      type: object
      description: Paginated list of movies.
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Movie'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
      required:
        - items
        - total
        - limit
        - offset
    Page_Series:
      type: object
      description: Paginated list of TV series.
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Series'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
      required:
        - items
        - total
        - limit
        - offset
    Page_Season:
      type: object
      description: Paginated list of seasons.
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Season'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
      required:
        - items
        - total
        - limit
        - offset
...
    Page_Stream:
      type: object
      description: Paginated list of video stream manifests.
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Stream'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
      required:
        - items
        - total
        - limit
        - offset
...
    Service:
      type: object
      description: |
        A streaming service platform (e.g. Netflix, Crave, Disney+).
        Maps to schema.org [Service](https://schema.org/Service).
      additionalProperties: false
      properties:
        '@type':
          type: string
          description: JSON-LD type discriminator.
          enum:
            - Service
          default: Service
        tag:
          type: string
          description: Short, unique service identifier.
          pattern: '^[A-Z]{2,8}$'
          examples:
            - CRAV
        name:
          type: string
          description: Human-readable service name.
          examples:
            - Crave
        description:
          type: string
          description: Short description of the service.
        website:
          type: string
          format: uri
          description: Official website URL.
          examples:
            - https://www.cravetv.ca
        country:
          type: string
          description: |
            ISO 3166-1 alpha-2 country code of the service's primary market.
          pattern: '^[A-Z]{2}$'
          examples:
            - CA
        languages:
          type: array
          description: |
            ISO 639-1 language codes available on this service.
          items:
            type: string
            pattern: '^[a-z]{2}$'
          examples:
            - en
            - fr
        logo:
          type: string
          format: uri
          description: URL to the service's logo image.
      required:
        - tag
        - name
    Movie:
      type: object
      description: |
        A self-contained film with no seasonal divisions.
        Maps to schema.org [Movie](https://schema.org/Movie).
      additionalProperties: false
      properties:
        '@type':
          type: string
          description: JSON-LD type discriminator.
          enum:
            - Movie
          default: Movie
        id:
          type: string
          description: Canonical API identifier (path-safe slug).
          examples:
            - inception-2010
        name:
          type: string
          description: Title of the movie.
          examples:
          - Inception
        description:
          type: string
          description: Plot summary or tagline.
        datePublished:
          type: string
          format: date
          description: |
            Original release date (ISO 8601 / RFC 3339 full-date).
          examples:
            - 2010-07-16
        duration:
          type: string
          description: |
            Runtime in ISO 8601 duration format.
          pattern: '^P(?!$)(?:\d+Y)?(?:\d+M)?(?:\d+D)?(?:T(?:\d+H)?(?:\d+M)?(?:\d+S)?)?$'
          examples:
            - PT2H28M
        countryOfOrigin:
          type: string
          description: |
            ISO 3166-1 alpha-2 country code of origin.
          pattern: '^[A-Z]{2}$'
          examples:
            - US
        languages:
          type: array
          description: |
            ISO 639-1 language codes available in this movie's audio or subtitles.
          items:
            type: string
            pattern: '^[a-z]{2}$'
          examples:
            - en
            - fr
        contentRating:
          type: string
          description: |
            Content rating from the source country's rating system
            (e.g. MPAA, TV Parental Guidelines).
          examples:
            - PG-13
        genres:
          type: array
          description: One or more genre classifications.
          items:
            type: string
          examples:
            - Action
            - Sci-Fi
            - Thriller
        poster:
          type: string
          format: uri
          description: URL to the movie poster image.
        backdrop:
          type: string
          format: uri
          description: URL to the movie backdrop image.
        trailer:
          type: string
          format: uri
          description: URL to a trailer video.
      required:
        - '@type'
        - id
        - name
...